refactor: split standard implementation into its own file
